feat(logger): Allow adding a custom name to go along with logging
This allows importers to pass a config property (`{ name:
'desiredName'}`), which will be used when logging all output.

The purpose of this is to distinguish which logger output lines come
from which importers. The output will now look like this, with
appropriate importer changes:

```
2019-02-22T01:21:11.484Z - info: [dbclient-openstreetmap]  paused=false, transient=0, current_length=171, indexed=17500, batch_ok=35, batch_retries=0, failed_records=0, venue=16387, address=1113, persec=1150
2019-02-22T01:21:21.483Z - info: [dbclient-openstreetmap]  paused=false, transient=1, current_length=180, indexed=34000, batch_ok=68, batch_retries=0, failed_records=0, venue=27652, address=6348, persec=1650
```

Fixes https://github.com/pelias/dbclient/issues/101